contract-projects/requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.4
cryptography==42.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0